jenkins_publish/requirements.txt
ldap3==2.5.1
python-jenkins==1.4.0
Flask==1.0.2